// src/api/generated/model/workoutSimpleResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */
import type { ProfileSimpleResponse } from './profileSimpleResponse';
import type { TagResponse } from './tagResponse';

export interface WorkoutSimpleResponse {
  id?: number;
  title?: string;
  author?: ProfileSimpleResponse;
  trainee?: ProfileSimpleResponse;
  tags?: TagResponse[];
  isTemplate?: boolean;
}